added tests for job_finished
